Separate advanced search expectations
